+++ include/quo-mod-env.h
@@ -40,13 +40,13 @@ extern "C" {
 
 // ---------- PRIVATE API ---------- //
 
-static inline QuoVar quo__mod_env_get(QuoModule *m, int64_t argc, QuoVar *argv) {
+static inline QuoVar quo__mod_env_get(QuoModule *m, int argc, QuoVar *argv) {
   if (argc != 1 || !quo_var_is_str(&argv[1])) return quo_var_new_err("env.get() requires a string argument");
   const char *value = getenv(quo_var_as_str(&argv[1])->data);
   return value ? quo_var_new_obj(quo_str_new(m, value, -1)) : quo_var_new_nil();
 }
 
-static inline QuoVar quo__mod_env_set(QuoModule *m, int64_t argc, QuoVar *argv) {
+static inline QuoVar quo__mod_env_set(QuoModule *m, int argc, QuoVar *argv) {
   if (argc != 2 || !quo_var_is_str(&argv[0]) || !quo_var_is_str(&argv[1]))
     return quo_var_new_err("env.set() requires key and value strings");
 #ifdef _WIN32
@@ -59,7 +59,7 @@ static inline QuoVar quo__mod_env_set(QuoModule *m, int64_t argc, QuoVar *argv) 
   return quo_var_new_nil();
 }
 
-static inline QuoVar quo__mod_env_unset(QuoModule *m, int64_t argc, QuoVar *argv) {
+static inline QuoVar quo__mod_env_unset(QuoModule *m, int argc, QuoVar *argv) {
   if (argc != 1 || !quo_var_is_str(&argv[0])) return quo_var_new_err("env.unset() requires a string argument");
 #ifdef _WIN32
   char *str = quo_strdupf("%s=", quo_var_as_str(&argv[0])->data);
@@ -71,12 +71,12 @@ static inline QuoVar quo__mod_env_unset(QuoModule *m, int64_t argc, QuoVar *argv
   return quo_var_new_nil();
 }
 
-static inline QuoVar quo__mod_env_has(QuoModule *m, int64_t argc, QuoVar *argv) {
+static inline QuoVar quo__mod_env_has(QuoModule *m, int argc, QuoVar *argv) {
   if (argc != 1 || !quo_var_is_str(&argv[0])) return quo_var_new_err("env.has() requires a string argument");
   return quo_var_new_bool(getenv(quo_var_as_str(&argv[0])->data) != NULL);
 }
 
-static inline QuoVar quo__mod_env_all(QuoModule *m, int64_t argc, QuoVar *argv) {
+static inline QuoVar quo__mod_env_all(QuoModule *m, int argc, QuoVar *argv) {
   QuoDict *dict = quo_dict_new();
 #ifdef _WIN32
   char *env_block = GetEnvironmentStrings();
